requirements.txt
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.2
cryptography==3.1
decorator==4.4.2
isort==5.5.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycparser==2.20
pylint==2.6.0
python-telegram-bot==12.8
six==1.15.0
toml==0.10.1
tornado==6.0.4
wrapt==1.12.1
